% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fec_version_info}
\alias{fec_version_info}
\title{Version of the bundled fec-parser crate}
\usage{
fec_version_info()
}
\value{
A string such as \code{"0.1.0"}.
}
\description{
Smoke test for the Rust bindings.
}
\examples{
fec_version_info()
}
